inline dbengine query critical path (#19537)
* inline dbengine query critical path

* more dbengine inlining

* optimize pgc_page_add to optimistically allocate the new entry

* backfill related inlining

* more inlining

* simple re-organization of members of sender_state

* descriptive SSL error logs

* nd-poll() wanted mismatch is now a debug statement

* improve SSL error handling

* restore ssl error handling code

* strict error checking on enable_streaming replication flag

* selected functions flattening

* size replication queue

* fix replication step inicfg parsing; fix local variable exposure

* replication sender code cleanup

* process opcodes every 100ms and reorganize so that the latency per processing can be computed

* more inlining at critical sections

* make replication step 20m

* configure replication prefetch

* use the inicfg range function for libuv threads

* move libuv initialize closer to the beginning

* replication fixes

* replication defaults tuning

* aral to use 2MiB page sizes and nd_mmap() to madvise HUGEPAGE when size matches

* use dedicated aral for replication sender

* tune replication sender to max 6 threads

* aral incoming lock now tries all partitions

* health should be running together with replication

* refcount added refcount_release_and_acquire_for_deletion_advanced()

* minimize aral_lock() use

* aral having page_lock

* aral_page refcount with simple atomics

* cleanup aral

+++ src/daemon/pulse/pulse-db-dbengine.h
@@ -6,7 +6,7 @@
 #include "daemon/common.h"
 
 #if defined(PULSE_INTERNALS)
-extern size_t pulse_dbengine_total_memory;
+extern int64_t pulse_dbengine_total_memory;
 
 #if defined(ENABLE_DBENGINE)
 void pulse_dbengine_do(bool extended);
